Crate wave2text

Structs

Settings ğŸ”’

Format to store/read settings in JSON file.

Functions

confirm ğŸ”’

Prompts user to confirm action before proceeding.

edit_waveform ğŸ”’

User menu for manually editing the waveform.

get_pulse_shape ğŸ”’

Loads pulse shape from TXT file.

get_settings ğŸ”’

Loads saved settings from JSON file.

get_user_num ğŸ”’

Gets user input and returns a number if valid

main ğŸ”’

terminal_menu ğŸ”’

Brings up the menu and returns the input if valid.

wave_gen ğŸ”’

Constructs a new waveform in which the provided pulse repeats as specified, and appends it to the end of the existing waveform.